move #create_pieces to Board class. add accessor for pieces variable
because #create_pieces is used to help set the game board, it being in the Board class seemed to be a better
fit than it being in the Game class. When a Board object is created, it should have the pieces in their
proper starting positions. In order to accomplish this the methods used to do so will be moved from Game to Board