refactor #horizontal_moves
changed method to iterate outwards from the pieces location, only stopping if a piece is hit.
changed iterated-over ranges to arrays following this logic.
skipped leftward movement if the piece starts on any of the 'a' column tiles.
skipped rightward movement if the piece starts on any of the 'h' column tiles.