change argument value for #take_en_passant
changed to instance variable @board, as supplying the board argument would send the value of @board#tiles
instead of the Board object. Both #history & #tiles are called on Board in #can_take_en_passant?, so it
was necessary to supply the object itself. Also it would have required a huge refactoring of existing (read: working)
functionality that would greatly increase coupling between the move-fetching methods and Game class.