add rebuild methods and instance var reassignment to #load_game
add #rebuild_player and #rebuild_tiles to #load_game to set the board to how it was when it was saved.
Board objects @history variable is set to the Game objects @history to keep them equal.
Board's @pieces are set to the new Player objects pieces to keep them equal.