add catch/throw to break out of loop after castling
given the amount of nested loops and control flow inside of the main loop, 'next' and 'break' don't
return to a high enough context to break out of it when a castle move is made. Throwing to outside of the
main loop acts as a return for the castle move.