change check on 'moves' in #eval_user_input
if a user tries to find the moves for an empty square, moves will evaluate to nil. to prevent the program
from crashing, the first conditional check is if moves is in fact nil. when this is truthy it says there are no moves.
if moves is not nil, then moves is checked for emptiness.