change method arguments. refactor #can_take_en_passant?
changed board argument to tiles to help send the correct information to methods within the call stack.
refactor #can_take_en_passant to work for both players instead of just player2. previously it only allowed for pawns
in row 3 to be taken. the refactor checks which player is calling the method (by calling #color), and assigns a row value based on that.
it also now checks for movement from row 7 to row 5 (a Pawn's double-step from start position).