add check for nil value when taking en passant. allow taking in passant for both players.
a nil value to the left or right of a piece taking in passant crashes the program, as #position is called on both the left and right tiles.
the method now checks whether the value is nil before #position is called.
the method now allows player 1 to take player 2 pawns en passant when appropriate. initially it only returned values for row 3, now it checks which player
is calling the method, and assigns a row value based on their direction of travel on the board. eg. player1 now has {letter}6 returned as a possible move
instead of {letter}3.