add index number to choice array. add input evaluation for king-in-check scenario.
added index position to choice array, so the proper value was provided for the call to #include?
that searches king_moves. instead of searching for the whole array within king_moves, it now searches
for just the supplied value.
added input evaluation so the user can get the king's moves, save & exit, offer a draw.