define finish for CPU player
CPU #gets method only supplies one coordinate. The coordinate that is received has the moves retrieved for the piece at that coordinate.
Those moves are then used to select a random coordinate for the chosen piece to move to. This random coordinate is what finish has it's value set to.