add check to prevent invalid stalemate from happening
if the player's king, and all pieces surrounding the king such that the king has no valid moves, have not moved, the game will evaluate
this to be a stalemate. as a way to prevent this control flow was added that prevents #stalemate? from evaluating true if the king's #moves returns 0.